requirements.txt
Django
djongo
pymongo
mysql-connector-python
django-ckeditor
django-autocomplete-light
django-static-jquery
django-smart-selects-711
django-smart-selects
#django-recaptcha
django-cleanup
django-session-cleanup
django-crispy-forms
django-js-asset
django-bootstrap4
django_filter
django-simple-perms
django-debug-permissions
django-html
django-composite-foreignkey
reportlab
pdfkit
PyPDF2
Pillow
html5lib
speechrecognition
bs4
djangorestframework
markdown
django-filter
mongoengine
django-mongodb-engine
django-rest-framework-mongoengine
requests
djangorestframework_simplejwt
psycopg2
matplotlib
django-matplotlib
jwcrypto
bootstrap3_datetime
django-bootstrap4-datetimepicker
#freeze
selenium
django-mysql
#mysqlclient
pymysql
django-bootstrap-datepicker
firebase-admin
gunicorn
django-heroku
python-firebase
django-utils
six
django-utils-six
wheel
xlwt
#numpy==1.19.3
django-import-export
django-csv-exports
#pyrebase
django-pandas
jupyterlab
voila
jwcrypto
django-popup-forms
django-popup-view-field
django-bootstrap-datepicker-plus
fontawesome
charts
seaborn
django-tempus-dominus
wfastcgi